+++ VevaciousPlusPlus/include/PotentialEvaluation/ThermalFunctions.hpp
@@ -8,7 +8,8 @@
 #ifndef THERMALFUNCTIONS_HPP_
 #define THERMALFUNCTIONS_HPP_
 
-#include "CommonIncludes.hpp"
+#include <string>
+#include <sstream>
 
 namespace VevaciousPlusPlus
 {
@@ -130,7 +131,7 @@ namespace VevaciousPlusPlus
   ThermalFunctions::BosonMinusOneToMinusTwelve( double const squareRatio )
   {
     double scaledRatio( -10.0 * ( squareRatio + 1.0 ) );
-    size_t floorIndex( scaledRatio );
+    size_t floorIndex( static_cast< size_t >( scaledRatio ) );
     return ( bosonMinusOneToMinusTwelve[ floorIndex ]
              + ( 0.1 * ( scaledRatio - floorIndex )
                      * ( bosonMinusOneToMinusTwelve[ floorIndex + 1 ]
@@ -142,7 +143,7 @@ namespace VevaciousPlusPlus
   ThermalFunctions::BosonZeroToMinusOne( double const squareRatio )
   {
     double scaledRatio( -100.0 * squareRatio );
-    size_t floorIndex( scaledRatio );
+    size_t floorIndex( static_cast< size_t >( scaledRatio ) );
     return ( bosonZeroToMinusOne[ floorIndex ]
              + ( 0.01 * ( scaledRatio - floorIndex )
                       * ( bosonZeroToMinusOne[ floorIndex + 1 ]
@@ -154,7 +155,7 @@ namespace VevaciousPlusPlus
   ThermalFunctions::BosonZeroToPlusOne( double const squareRatio )
   {
     double scaledRatio( 100.0 * squareRatio );
-    size_t floorIndex( scaledRatio );
+    size_t floorIndex( static_cast< size_t >( scaledRatio ) );
     return ( bosonZeroToPlusOne[ floorIndex ]
              + ( 0.01 * ( scaledRatio - floorIndex )
                       * ( bosonZeroToPlusOne[ floorIndex + 1 ]
@@ -166,7 +167,7 @@ namespace VevaciousPlusPlus
   ThermalFunctions::BosonPlusOneToPlusOneHundred( double const squareRatio )
   {
     double scaledRatio( squareRatio - 1.0 );
-    size_t floorIndex( scaledRatio );
+    size_t floorIndex( static_cast< size_t >( scaledRatio ) );
     return ( bosonPlusOneToPlusOneHundred[ floorIndex ]
              + ( ( scaledRatio - floorIndex )
                  * ( bosonPlusOneToPlusOneHundred[ floorIndex + 1 ]
@@ -178,7 +179,7 @@ namespace VevaciousPlusPlus
   ThermalFunctions::FermionMinusOneToMinusTwelve( double const squareRatio )
   {
     double scaledRatio( -10.0 * ( squareRatio + 1.0 ) );
-    size_t floorIndex( scaledRatio );
+    size_t floorIndex( static_cast< size_t >( scaledRatio ) );
     return ( fermionMinusOneToMinusTwelve[ floorIndex ]
              + ( 0.1 * ( scaledRatio - floorIndex )
                      * ( fermionMinusOneToMinusTwelve[ floorIndex + 1 ]
@@ -190,7 +191,7 @@ namespace VevaciousPlusPlus
   ThermalFunctions::FermionZeroToMinusOne( double const squareRatio )
   {
     double scaledRatio( -100.0 * squareRatio );
-    size_t floorIndex( scaledRatio );
+    size_t floorIndex( static_cast< size_t >( scaledRatio ) );
     return ( fermionZeroToMinusOne[ floorIndex ]
              + ( 0.01 * ( scaledRatio - floorIndex )
                       * ( fermionZeroToMinusOne[ floorIndex + 1 ]
@@ -202,7 +203,7 @@ namespace VevaciousPlusPlus
   ThermalFunctions::FermionZeroToPlusOne( double const squareRatio )
   {
     double scaledRatio( 100.0 * squareRatio );
-    size_t floorIndex( scaledRatio );
+    size_t floorIndex( static_cast< size_t >( scaledRatio ) );
     return ( fermionZeroToPlusOne[ floorIndex ]
              + ( 0.01 * ( scaledRatio - floorIndex )
                       * ( fermionZeroToPlusOne[ floorIndex + 1 ]
@@ -214,7 +215,7 @@ namespace VevaciousPlusPlus
   ThermalFunctions::FermionPlusOneToPlusOneHundred( double const squareRatio )
   {
     double scaledRatio( squareRatio - 1.0 );
-    size_t floorIndex( scaledRatio );
+    size_t floorIndex( static_cast< size_t >( scaledRatio ) );
     return ( fermionPlusOneToPlusOneHundred[ floorIndex ]
              + ( ( scaledRatio - floorIndex )
                  * ( fermionPlusOneToPlusOneHundred[ floorIndex + 1 ]
